+++ azext_edge/edge/providers/rpsaas/user_strings.py
@@ -1,0 +1,19 @@
+# coding=utf-8
+# ----------------------------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License file in the project root for license information.
+# ----------------------------------------------------------------------------------------------
+
+# Base Strings
+CUSTOM_LOCATION_DOES_NOT_EXIST_ERROR = "Cluster associated with custom location {0} does not exist."
+CUSTOM_LOCATION_NOT_FOUND_MSG = "Custom location {0} not found. The command may fail."
+CLUSTER_NOT_FOUND_MSG = "Cluster associated with the custom location {0} not found. "\
+    "The command may fail."
+CLUSTER_OFFLINE_MSG = "Cluster {0} is not connected. The cluster may not update correctly."
+MISSING_CLUSTER_CUSTOM_LOCATION_ERROR = "Need to provide either cluster name or custom location"
+MISSING_EXTENSION_ERROR = "Cluster {0} is missing the {1} extension."
+MULTIPLE_CUSTOM_LOCATIONS_ERROR = "The following custom locations were found for cluster {0}: \n{1}. "\
+    "Please specify which custom location to use."
+MULTIPLE_POSSIBLE_ITEMS_ERROR = "Found {0} {1}s with the name {2}. Please provide the resource group "\
+    "for the {1}."
+NO_EXTENDED_LOCATION_TO_CHECK_MSG = "No extended location is associated. Cluster checks are skipped."
