+++ azext_edge/tests/edge/rpsaas/adr/asset_profile/test_transport_auth_unit.py
@@ -14,7 +14,7 @@ from azext_edge.edge.commands_asset_endpoint_profiles import (
 from azext_edge.edge.providers.rpsaas.adr.base import ADR_API_VERSION
 
 from .conftest import MINIMUM_AEP, FULL_AEP
-from ..conftest import RM_PATH
+from ...conftest import BP_PATH
 from .....generators import generate_generic_id
 
 
@@ -22,7 +22,7 @@ FULL_CERT = FULL_AEP["properties"]["transportAuthentication"]["ownCertificates"]
 
 
 @pytest.mark.parametrize("mocked_build_query", [{
-    "path": RM_PATH,
+    "path": BP_PATH,
     "result": [{"properties": {"connectivityStatus": "Connected"}}]
 }], ids=["query"], indirect=True)
 @pytest.mark.parametrize("mocked_resource_management_client", [
@@ -99,7 +99,7 @@ def test_list_asset_endpoint_profile_transport_auths(mocked_cmd, mocked_resource
 
 
 @pytest.mark.parametrize("mocked_build_query", [{
-    "path": RM_PATH,
+    "path": BP_PATH,
     "result": [{"properties": {"connectivityStatus": "Connected"}}]
 }], ids=["query"], indirect=True)
 @pytest.mark.parametrize("mocked_resource_management_client", [
