+++ azext_edge/tests/edge/rpsaas/base/provider/test_show_unit.py
@@ -11,11 +11,17 @@ context_name = generate_generic_id()
 
 
 @pytest.mark.parametrize("mocked_resource_management_client", [
-    {"resources.get": {"extended_location": generate_generic_id()}},
-    {"resources.get": {"extendedLocation": generate_generic_id()}},
+    {"resources.get": {"extended_location": {"name": generate_generic_id()}}},
+    {"resources.get": {"extendedLocation": {"name": generate_generic_id()}}},
     {"resources.get": {"result": generate_generic_id()}},
 ], ids=["extended_location", "extendedLocation", "result"], indirect=True)
-def test_show(mocked_cmd, mocked_resource_management_client):
+@pytest.mark.parametrize("check_cluster_connectivity", [True, False])
+def test_show(
+    mocked_cmd,
+    mock_check_cluster_connectivity,
+    mocked_resource_management_client,
+    check_cluster_connectivity
+):
     from azext_edge.edge.providers.rpsaas.base_provider import RPSaaSBaseProvider
     api_version = generate_generic_id()
     resource_type = generate_generic_id()
@@ -26,7 +32,7 @@ def test_show(mocked_cmd, mocked_resource_management_client):
         api_version,
         resource_type
     )
-    result = provider.show(resource_name, resource_group_name)
+    result = provider.show(resource_name, resource_group_name, check_cluster_connectivity)
 
     mocked_resource_management_client.resources.get.assert_called_once()
     kwargs = mocked_resource_management_client.resources.get.call_args.kwargs
@@ -40,3 +46,9 @@ def test_show(mocked_cmd, mocked_resource_management_client):
     as_dict = mocked_resource_management_client.resources.get.return_value.as_dict.return_value
     assert "extended_location" not in result
     assert result == as_dict
+
+    assert mock_check_cluster_connectivity.called is check_cluster_connectivity
+    if check_cluster_connectivity:
+        mock_check_cluster_connectivity.assert_called_once_with(
+            as_dict.get("extendedLocation", {}).get("name")
+        )
